final reve

+++ src/utils/axios.js
@@ -32,9 +32,20 @@ axiosApiIntances.interceptors.response.use(
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
     if (error.response.status === 403) {
-      localStorage.clear();
-      alert("Please log in with a verified account !");
-      window.location.href = "/login";
+      if (error.response.data.msg === "jwt expired") {
+        const refreshToken = localStorage.getItem("refreshToken");
+        axiosApiIntances
+          .post("auth/refresh", { refreshToken })
+          .then((res) => {
+            localStorage.setItem("token", res.data.data.token);
+            window.location.reload();
+          })
+          .catch((err) => console.log(err));
+      } else {
+        localStorage.clear();
+        alert("Please log in with a verified account !");
+        window.location.href = "/login";
+      }
     }
     return Promise.reject(error);
   }
